New SqlAzureServerClient.cs
Adding a SqlAzureServerClient to override some of the methods for
dropping and creating new databases. There are a few steps in
SQL Azure that can't be done together in T-SQL and need to be split
apart. Also, the DB_ID('') function on SQL Azure always returns NULL
if called from the master database for another database name.

Also, updated the Sample program to be a little more verbose and
use a session when getting a count since the ToEnumerableAsJson
method doesn't seem to be implemented in SisoReadOnce.